Fix bug in upload where source import files could be removed if an error occurs during detection and galaxy owns the source files.

+++ tools/data_source/upload.py
@@ -24,6 +24,9 @@ def file_err( msg, dataset, json_file ):
                                            ext = 'data',
                                            dataset_id = dataset.dataset_id,
                                            stderr = msg ) ) + "\n" )
+    # never remove a server-side upload
+    if dataset.type in ( 'server_dir', 'path_paste' ):
+        return
     try:
         os.remove( dataset.path )
     except:
